+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+
+## 3.1.0
+
+- Added support for passing custom HTTP clients with global headers (e.g., User-Agent) via `AppClient` and `InterceptedClient`.
+- Improved HTTP client abstraction to allow using either `http.Client` or custom clients.
+- Enhanced documentation for helper classes and HTTP client usage.
+- Improved error handling and logging in network helpers.
+- added flutter side example [flutter_example]
+- improve the stracture of the package
+- improve README.MD
+
+
 ## 3.0.1
 - Fixed zoom attribute